@@ -77,8 +77,11 @@
     "electron-webpack": "^2.8.2",
     "enzyme-adapter-react-16": "^1.15.4",
     "enzyme-to-json": "^3.5.0",
-    "eslint": "^7.9.0",
-    "eslint-plugin-react": "^7.20.6",
+    "eslint": "^7.28.0",
+    "eslint-plugin-import": "^2.23.4",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-react": "^7.24.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "nodemon": "^2.0.7",
     "react-test-renderer": "^16.13.1",
     "redux-thunk": "^2.3.0",
