@@ -1,22 +1,19 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import store from "./store";
-import { Provider } from "react-redux";
-import App from "../components/App";
-import "../components/css/styles.css";
-import "../components/css/metric.css";
-import "../components/css/running.css";
-import "../components/css/static.css";
-import Login from "../components/login/login";
-import { HashRouter as Router } from "react-router-dom";
-
-//const firebase = require('firebase');
-// const firebaseui = require('firebaseui');
+import React from 'react';
+import ReactDOM from 'react-dom';
+import store from './store';
+import { Provider } from 'react-redux';
+import App from '../components/App';
+import '../components/css/styles.css';
+import '../components/css/metric.css';
+import '../components/css/running.css';
+import '../components/css/static.css';
+import Login from '../components/login/login';
+import { HashRouter as Router } from 'react-router-dom';
 /**
  * Electron's path configuration is done invoking fixPath
  * Will not run without this logic
  */
-const fixPath = require("fix-path");
+const fixPath = require('fix-path');
 fixPath();
 
 ReactDOM.render(
@@ -25,6 +22,6 @@ ReactDOM.render(
       <Login />
     </Router>
   </Provider>,
-  document.getElementById("app")
+  document.getElementById('app')
 
 );
