@@ -1,39 +1,49 @@
 module.exports = {
-    "env": {
-        "browser": true,
-        "es2021": true,
-        "node": true
+  'env': {
+    'browser': true,
+    'commonjs': true,
+    'es2021': true,
+    'node': true
+  },
+  'extends': [
+    'eslint:recommended',
+    'plugin:react/recommended'
+  ],
+  'parserOptions': {
+    'ecmaFeatures': {
+      'jsx': true,
     },
-    "extends": "plugin:react/recommended",
-    "parserOptions": {
-        "ecmaFeatures": {
-            "jsx": true
-        },
-        "ecmaVersion": 12,
-        "sourceType": "module"
-    },
-    "plugins": [
-        "react"
+    'ecmaVersion': 12,
+    'sourceType': 'module'
+  },
+  'plugins': [
+    'react'
+  ],
+  'rules': {
+    'indent': ['warn', 2],
+    'no-unused-vars': ['off', { vars: 'local' }],
+    'prefer-const': 'warn',
+    quotes: ['warn', 'single'],
+    semi: ['warn', 'always'],
+    'space-infix-ops': 'warn',
+    'no-console': 'off',
+    'no-restricted-syntax': [
+      'error',
+      {
+        'selector': 'CallExpression[callee.object.name=\'console\'][callee.property.name!=/^(log|warn|error|info|trace)$/]',
+        'message': 'Unexpected property on console object was called'
+      }
     ],
-    "rules": {
-        "no-console": "off",
-        "no-restricted-syntax": [
-            "error",
-            {
-                "selector": "CallExpression[callee.object.name='console'][callee.property.name!=/^(log|warn|error|info|trace)$/]",
-                "message": "Unexpected property on console object was called"
-            }
-        ],
-        "spaced-comment": ["error", "always", {
-            "line": {
-                "markers": ["/"],
-                "exceptions": ["-", "+"]
-            },
-            "block": {
-                "markers": ["!"],
-                "exceptions": ["*"],
-                "balanced": true
-            }
-        }]
-    }
+    'spaced-comment': ['error', 'always', {
+      'line': {
+        'markers': ['/'],
+        'exceptions': ['-', '+']
+      },
+      'block': {
+        'markers': ['!'],
+        'exceptions': ['*'],
+        'balanced': true
+      }
+    }]
+  }
 };
