@@ -11,7 +11,7 @@ router.post('/',
   (req, res) => {
     console.log('made it back to login router');
     if (res.locals.error) return res.status(200).json(res.locals);
-    return res.status(200).json(res.locals.id);
+    return res.status(200).json(res.locals);
   }
 );
 
