@@ -1,4 +1,4 @@
-import * as types from "../constants/actionTypes";
+import * as types from '../constants/actionTypes';
 
 export const addRunningContainers = (data) => ({
   type: types.ADD_RUNNING_CONTAINERS,
@@ -137,5 +137,19 @@ export const addNotificationFrequency = (data) => ({
 
 export const addMonitoringFrequency = (data) => ({
   type: types.MONITORING_FREQUENCY,
+  payload: data,
+});
+
+export const updateSession = () => ({
+  type: types.UPDATE_SESSION,
+});
+
+export const updateUser = (data) => ({
+  type: types.UPDATE_USER,
+  payload: data,
+});
+
+export const logoutUser = (data) => ({
+  type: types.LOGOUT_USER,
   payload: data,
 });
