@@ -1,36 +1,31 @@
-export const ADD_RUNNING_CONTAINERS = "ADD_RUNNING_CONTAINERS";
-export const REMOVE_CONTAINER = "REMOVE_CONTAINER";
-export const STOP_RUNNING_CONTAINER = `STOP_RUNNING_CONTAINER`;
-export const ADD_STOPPED_CONTAINERS = `ADD_STOPPED_CONTAINERS`;
-export const RUN_STOPPED_CONTAINER = `RUN_STOPPED_CONTAINER`;
-export const GET_IMAGES = "GET_IMAGES";
-export const RUN_IMAGE = "RUN_IMAGE";
-export const REMOVE_IMAGE = "REMOVE_IMAGE";
-export const REFRESH_RUNNING_CONTAINERS = "REFRESH_RUNNING_CONTAINERS";
-export const REFRESH_STOPPED_CONTAINERS = "REFRESH_STOPPED_CONTAINERS";
-export const REFRESH_IMAGES = "REFRESH_IMAGES";
-export const COMPOSE_YML_FILES = "COMPOSE_YML_FILES";
-export const GET_NETWORK_CONTAINERS = "GET_NETWORK_CONTAINERS";
-export const GET_CONTAINER_STACKS = "GET_CONTAINER_STACKS";
-export const COMPOSE_DOWN = "COMPOSE_DOWN";
-export const COMPOSE_UP = "COMPOSE_UP";
-export const BUILD_AXIS = "BUILD_AXIS";
-export const BUILD_MEMORY = "BUILD_MEMORY";
-export const BUILD_CPU = "BUILD_CPU";
-export const ADD_PHONE_NUMBER = "ADD_PHONE_NUMBER";
-export const ADD_MEMORY_NOTIFICATION_SETTING =
-  "ADD_MEMORY_NOTIFICATION_SETTING";
-export const ADD_CPU_NOTIFICATION_SETTING = "ADD_CPU_NOTIFICATION_SETTING";
-export const ADD_STOPPED_NOTIFICATION_SETTING =
-  "ADD_STOPPED_NOTIFICATION_SETTING";
-export const REMOVE_MEMORY_NOTIFICATION_SETTING =
-  "REMOVE_MEMORY_NOTIFICATION_SETTING";
-export const REMOVE_CPU_NOTIFICATION_SETTING =
-  "REMOVE_CPU_NOTIFICATION_SETTING";
-export const REMOVE_STOPPED_NOTIFICATION_SETTING =
-  "REMOVE_STOPPED_NOTIFICATION_SETTING";
-export const NOTIFICATION_FREQUENCY = "NOTIFICATION_FREQUENCY";
-export const MONITORING_FREQUENCY = "MONITORING_FREQUENCY";
+export const ADD_RUNNING_CONTAINERS = 'ADD_RUNNING_CONTAINERS';
+export const REMOVE_CONTAINER = 'REMOVE_CONTAINER';
+export const STOP_RUNNING_CONTAINER = 'STOP_RUNNING_CONTAINER';
+export const ADD_STOPPED_CONTAINERS = 'ADD_STOPPED_CONTAINERS';
+export const RUN_STOPPED_CONTAINER = 'RUN_STOPPED_CONTAINER';
+export const GET_IMAGES = 'GET_IMAGES';
+export const RUN_IMAGE = 'RUN_IMAGE';
+export const REMOVE_IMAGE = 'REMOVE_IMAGE';
+export const REFRESH_RUNNING_CONTAINERS = 'REFRESH_RUNNING_CONTAINERS';
+export const REFRESH_STOPPED_CONTAINERS = 'REFRESH_STOPPED_CONTAINERS';
+export const REFRESH_IMAGES = 'REFRESH_IMAGES';
+export const COMPOSE_YML_FILES = 'COMPOSE_YML_FILES';
+export const GET_NETWORK_CONTAINERS = 'GET_NETWORK_CONTAINERS';
+export const GET_CONTAINER_STACKS = 'GET_CONTAINER_STACKS';
+export const COMPOSE_DOWN = 'COMPOSE_DOWN';
+export const COMPOSE_UP = 'COMPOSE_UP';
+export const BUILD_AXIS = 'BUILD_AXIS';
+export const BUILD_MEMORY = 'BUILD_MEMORY';
+export const BUILD_CPU = 'BUILD_CPU';
+export const ADD_PHONE_NUMBER = 'ADD_PHONE_NUMBER';
+export const ADD_MEMORY_NOTIFICATION_SETTING = 'ADD_MEMORY_NOTIFICATION_SETTING';
+export const ADD_CPU_NOTIFICATION_SETTING = 'ADD_CPU_NOTIFICATION_SETTING';
+export const ADD_STOPPED_NOTIFICATION_SETTING = 'ADD_STOPPED_NOTIFICATION_SETTING';
+export const REMOVE_MEMORY_NOTIFICATION_SETTING = 'REMOVE_MEMORY_NOTIFICATION_SETTING';
+export const REMOVE_CPU_NOTIFICATION_SETTING = 'REMOVE_CPU_NOTIFICATION_SETTING';
+export const REMOVE_STOPPED_NOTIFICATION_SETTING = 'REMOVE_STOPPED_NOTIFICATION_SETTING';
+export const NOTIFICATION_FREQUENCY = 'NOTIFICATION_FREQUENCY';
+export const MONITORING_FREQUENCY = 'MONITORING_FREQUENCY';
 export const UPDATE_NAME = 'UPDATE_NAME';
 export const UPDATE_EMAIL = 'UPDATE_EMAIL';
 export const UPDATE_PHONE = 'UPDATE_PHONE';
@@ -40,3 +35,8 @@ export const UPDATE_MEM_THRESHOLD = 'UPDATE_MEM_THRESHOLD';
 export const UPDATE_CPU_THRESHOLD = 'UPDATE_CPU_THRESHOLD';
 export const UPDATE_CONTAINER_STOPS = 'UPDATE_CONTAINER_STOPS';
 export const UPDATE_ALL = 'UPDATE_ALL';
+
+
+export const UPDATE_SESSION = 'UPDATE_SESSION';
+export const UPDATE_USER = 'UPDATE_USER';
+export const LOGOUT_USER = 'LOGOUT_USER';
