@@ -1,9 +1,10 @@
-import { combineReducers } from "redux";
-import graphsReducer from "./graphReducer";
-import imageListReducer from "./imageListReducer";
-import containerListReducer from "./containerListReducer";
-import dockerComposeReducer from "./dockerComposeReducer";
-import notificationReducer from "./notificationReducer";
+import { combineReducers } from 'redux';
+import graphsReducer from './graphReducer';
+import imageListReducer from './imageListReducer';
+import containerListReducer from './containerListReducer';
+import dockerComposeReducer from './dockerComposeReducer';
+import notificationReducer from './notificationReducer';
+import sessionReducer from './sessionReducer';
 
 /**
  * Reducer store
@@ -14,6 +15,7 @@ const reducers = combineReducers({
   containersList: containerListReducer,
   networkList: dockerComposeReducer,
   notificationList: notificationReducer,
+  session: sessionReducer,
 });
 
 export default reducers;
